Componentize export
